adding resample function to surface.
In the current implementation, resample fails if the matrix for the
given order doesn't exist. Despite this, the interface is setup to be
general in the hope that we will fix the matrix headache in the
future.

Currently there is only one upsample matrix that can be used for
interaction or reparametrization.

Adding resample needed storing shorder in the surface class. and
propagating the change in signature to other files and tests.

Also cleaned up the parameters so that it's clear what filtering
frequency is used (two filtering frequencies are in the code one is
for differentiation and the other is for reparametrization).