Update reparameterization
When we detect that the mesh is bad, we move points around so that they
are approximately equispaced (on the upsampled mesh).  This will
introduce high frequencies, but we should be fine since we upsampled.
Then we use the original reparameterization code, so that when we
truncate the truncation error is small.

The assumption above is that the required change in parameterization is
either orthogonal to the change that original reparameterization makes,
or the original reparameterization gets stuck in some local minima where
the mesh is bad.  So it needs to be nudged to a good parameterization in
the upsampled mesh and then original reparameterization continues after
that.