Making the reparametrization type an argument.
Also exposing the attenuation coefficient exponent for testing. Now
there are two types of reparametrization Box and PolyK. Box is the old
method with box filter. PolyK is one with 1-(n/n*)^k (for n<n*)
attenuation coefficient. In case of PolyK, k is a float and can be
smaller or greater than one. Large k gives a filter very similar to
the Box case.