Fix an issue which caused catastrophic cancellation.
During far-field evaluation, when surface points are not filtered (i.e. not in
Spherical Harmonic space), it results in some source and target points which do
not exactly coincide and are not recognized by the kernel function as the same
point (r=0 case). The potential at such target points blows up.

The current fix solves this issue by also filtering the target points and
increasing the tolerance in the single and double layer Stokes kernel functions
(i.e. if r<tol then the potential from such source points is not computed).